--- neo_tests/base_config.yaml
@@ -1,48 +1,0 @@
-global:
-  developer_mode: False
-
-rest:
-  verify: False
-  disable_ssl_warnings: True
-
-logging:
-  default_log_filename: 'pyvcloud-test.log'
-  log_requests: True
-  log_headers: True
-  log_bodies: True
-
-vcd:
-  host: '10.150.198.193'
-  api_version: '29.0'
-  sys_admin_username: 'administrator'
-  sys_admin_pass: 'ca$hc0w'
-  sys_org_name: 'System'
-
-  default_pvdc_name: '*'
-
-  default_org_name: 'test-org'
-
-  default_org_user_password : 'vmware'
-
-  default_ovdc_name: 'test-org-ovdc'
-  default_storage_profile_name: '*'
-  default_netpool_name: '*'
-
-  default_catalog_name: 'test-cat'
-  default_template_file_name: 'test_vapp_template.ova'
-
-  default_vapp_name: 'test_vapp'
-  default_vm_name: 'testvm1'
-
-vc:
-  vcenter_host_name: 'vc1'
-  vcenter_host_ip: ''
-  vcenter_admin_username: ''
-  vcenter_admin_password: ''
-
-nsx:
-  nsx_hostname: ''
-  nsx_host_ip: ''
-  nsx_admin_username: ''
-  nsx_admin_password: ''
-
