+++ system_tests/base_config.yaml
@@ -1,0 +1,67 @@
+# Config parameters for running system tests on pyvcloud
+
+global:
+# If this is turned on, any method decorated with @developer_mode will not be 
+# executed by unittest. Primarily used to suppress running teardown test 
+# methods.
+  developer_mode: False
+
+connection:
+# Supress warnings generated by unittest related to unclosed sockets, 
+# leaking http connections or missing https certificates
+  verify: False
+  disable_ssl_warnings: True
+
+logging:
+  default_log_filename: 'pyvcloud-system-test.log'
+  log_requests: True
+  log_headers: True
+  log_bodies: True
+
+
+vcd:
+# Fill in the vCloud Dirctor host, sysadmin credentials and other related 
+# information here. These will be required for the tests to run.
+  host: '10.150.198.193'
+  api_version: '29.0'
+  sys_admin_username: 'administrator'
+  sys_admin_pass: 'ca$hc0w'
+  sys_org_name: 'System'
+
+# The following params defines the org, pvdc, ovdc, etc. that the tests will 
+# run in / use. It's better to not change them.
+
+# * = choose first available pdvc
+  default_pvdc_name: '*'
+
+  default_org_name: 'pyvcloud-stystem-test-org'
+
+# org_admin, catalog_author, vapp_author, vapp_user and console_user are 
+# created by default, their password will be dictated by the following
+  default_org_user_password : 'vmware'
+
+  default_ovdc_name: 'pyvcloud-system-test-org-vdc'
+# * = choose first available storage profile
+  default_storage_profile_name: '*'
+# * = choose first available netpool
+  default_netpool_name: '*'
+
+  default_catalog_name: 'test-cat'
+  default_template_file_name: 'test_vapp_template.ova'
+
+  default_vapp_name: 'test_vapp'
+  default_vm_name: 'testvm1'
+
+# vCenter and NSX that must be attached to vCD prior to running tests
+vc:
+  vcenter_host_name: 'vc1'
+  vcenter_host_ip: ''
+  vcenter_admin_username: ''
+  vcenter_admin_password: ''
+
+nsx:
+  nsx_hostname: 'nsx1'
+  nsx_host_ip: ''
+  nsx_admin_username: ''
+  nsx_admin_password: ''
+
